+++ apps/microsoft-teams/frontend/src/components/config/ContentTypeSelectionModal/ContentTypeSelectionModal.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useState, useCallback } from 'react';
 import { Button, FormControl, Modal, Radio, Table } from '@contentful/f36-components';
 import { contentTypeSelection } from '@constants/configCopy';
 import { styles } from './ContentTypeSelectionModal.styles';
@@ -7,7 +7,9 @@ import ModalHeader from '@components/config/ModalHeader/ModalHeader';
 import { ContentTypeProps } from 'contentful-management';
 import EmptyState from '@components/config/EmptyState/EmptyState';
 import WebApp from '@components/config/EmptyState/WebApp';
+import DebouncedSearchInput from '@components/config/DebouncedSearchInput/DebouncedSearchInput';
 import ErrorMessage from '../ErrorMessage/ErrorMessage';
+import SearchableList from '@components/config/SearchableList/SearchableList';
 
 interface Props {
   isShown: boolean;
@@ -31,9 +33,32 @@ const ContentTypeSelectionModal = (props: Props) => {
   } = props;
 
   const [selectedContentTypeId, setSelectedContentTypeId] = useState(savedContentTypeId ?? '');
+  const [searchQuery, setSearchQuery] = useState<string>('');
 
-  const { title, button, link, emptyContent, emptyHeading, errorMessage } =
+  const { title, button, link, emptyContent, emptyHeading, errorMessage, searchPlaceholder } =
     contentTypeSelection.modal;
+
+  const handleSearchQueryUpdate = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setSearchQuery(e.target.value);
+  };
+
+  const renderRow = useCallback(
+    (contentType: ContentTypeProps) => {
+      return (
+        <Table.Row
+          key={contentType.sys.id}
+          onClick={() => setSelectedContentTypeId(contentType.sys.id)}
+          className={styles.tableRow}>
+          <Table.Cell>
+            <Radio id={contentType.sys.id} isChecked={selectedContentTypeId === contentType.sys.id}>
+              {contentType.name}
+            </Radio>
+          </Table.Cell>
+        </Table.Row>
+      );
+    },
+    [selectedContentTypeId]
+  );
 
   const renderModalContent = () => {
     if (error) {
@@ -49,25 +74,18 @@ const ContentTypeSelectionModal = (props: Props) => {
         <>
           <Modal.Content>
             <FormControl as="fieldset" marginBottom="none">
+              <DebouncedSearchInput
+                placeholder={searchPlaceholder}
+                onChange={handleSearchQueryUpdate}
+              />
               <Table className={styles.table}>
                 <Table.Body>
-                  {contentTypes.map((contentType) => (
-                    <Table.Row
-                      key={contentType.sys.id}
-                      onClick={() => setSelectedContentTypeId(contentType.sys.id)}
-                      className={styles.tableRow}>
-                      <Table.Cell>
-                        <Radio
-                          id={contentType.sys.id}
-                          isChecked={selectedContentTypeId === contentType.sys.id}
-                          onChange={() => {
-                            /* clicking entire row checks this radio, i.e. do nothing here */
-                          }}>
-                          {contentType.name}
-                        </Radio>
-                      </Table.Cell>
-                    </Table.Row>
-                  ))}
+                  <SearchableList
+                    list={contentTypes}
+                    searchQuery={searchQuery}
+                    renderListItem={renderRow}
+                    searchKeys={['sys.id', 'displayField', 'name', 'description']}
+                  />
                 </Table.Body>
               </Table>
             </FormControl>
@@ -90,6 +108,7 @@ const ContentTypeSelectionModal = (props: Props) => {
 
     return (
       <Modal.Content>
+        <DebouncedSearchInput placeholder={searchPlaceholder} disabled={true} />
         <EmptyState
           image={<WebApp />}
           heading={emptyHeading}